Handle DocBlock Resolving as separate driver + handle single types in doc blocks
