+++ tests/Fixtures/DocBlockType/Collection/CollectionOfClassesFromDifferentNamespace.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace JMS\Serializer\Tests\Fixtures\TypedProperties\Collection;
+namespace JMS\Serializer\Tests\Fixtures\DocBlockType\Collection;
 
-use JMS\Serializer\Tests\Fixtures\TypedProperties\Collection\Details\ProductDescription;
+use JMS\Serializer\Tests\Fixtures\DocBlockType\Collection\Details\ProductDescription;
 
 class CollectionOfClassesFromDifferentNamespace
 {
