+++ tests/Fixtures/DocBlockType/Collection/CollectionOfClassesFromSameNamespace.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace JMS\Serializer\Tests\Fixtures\TypedProperties\Collection;
+namespace JMS\Serializer\Tests\Fixtures\DocBlockType\Collection;
 
 class CollectionOfClassesFromSameNamespace
 {
