+++ tests/Fixtures/DocBlockType/Collection/CollectionOfClassesWithFullNamespacePath.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JMS\Serializer\Tests\Fixtures\DocBlockType\Collection;
+
+class CollectionOfClassesWithFullNamespacePath
+{
+    /**
+     * @var JMS\Serializer\Tests\Fixtures\DocBlockType\Collection\Product[]
+     */
+    public array $productIds;
+}
