+++ tests/Fixtures/DocBlockType/Collection/Details/ProductName.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace JMS\Serializer\Tests\Fixtures\TypedProperties\Collection\Details;
+namespace JMS\Serializer\Tests\Fixtures\DocBlockType\Collection\Details;
 
 class ProductName
 {
