+++ tests/Fixtures/DocBlockType/Collection/Details/WithProductDescriptionTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace JMS\Serializer\Tests\Fixtures\TypedProperties\Collection\Details;
+namespace JMS\Serializer\Tests\Fixtures\DocBlockType\Collection\Details;
 
 trait WithProductDescriptionTrait
 {
