+++ tests/Fixtures/DocBlockType/Collection/WithTraitInsideTrait.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JMS\Serializer\Tests\Fixtures\DocBlockType\Collection;
+
+use JMS\Serializer\Tests\Fixtures\DocBlockType\Collection\Details\WithProductDescriptionTrait;
+
+trait WithTraitInsideTrait
+{
+    use WithProductDescriptionTrait;
+}
