+++ tests/Fixtures/DocBlockType/SingleClassFromDifferentNamespaceTypeHint.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace JMS\Serializer\Tests\Fixtures\DocBlockType;
+
+use JMS\Serializer\Tests\Fixtures\DocBlockType\Collection\Details\ProductDescription;
+
+class SingleClassFromDifferentNamespaceTypeHint
+{
+    /**
+     * @var ProductDescription
+     */
+    private $data;
+}
