+++ tests/Fixtures/DocBlockType/SingleClassFromGlobalNamespaceTypeHint.php
@@ -1,0 +1,13 @@
+<?php
+
+
+namespace JMS\Serializer\Tests\Fixtures\DocBlockType;
+
+
+class SingleClassFromGlobalNamespaceTypeHint
+{
+    /**
+     * @var \stdClass
+     */
+    private $data;
+}
