--- tests/Fixtures/TypedProperties/Collection/CollectionOfClassesFromTrait.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace JMS\Serializer\Tests\Fixtures\TypedProperties\Collection;
-
-use JMS\Serializer\Tests\Fixtures\TypedProperties\Collection\Details\WithProductDescriptionTrait;
-
-class CollectionOfClassesFromTrait
-{
-    use WithProductDescriptionTrait;
-    use WithProductNameTrait;
-}
