--- tests/Fixtures/TypedProperties/Collection/CollectionOfClassesWithFullNamespacePath.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace JMS\Serializer\Tests\Fixtures\TypedProperties\Collection;
-
-class CollectionOfClassesWithFullNamespacePath
-{
-    /**
-     * @var JMS\Serializer\Tests\Fixtures\TypedProperties\Collection\Product[]
-     */
-    public array $productIds;
-}
