--- tests/Fixtures/TypedProperties/Collection/IncorrectCollection.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace JMS\Serializer\Tests\Fixtures\TypedProperties\Collection;
-
-class IncorrectCollection
-{
-    /**
-     * @var \stdClass
-     */
-    public array $productIds;
-}
