--- tests/Fixtures/TypedProperties/Collection/WithTraitInsideTrait.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace JMS\Serializer\Tests\Fixtures\TypedProperties\Collection;
-
-use JMS\Serializer\Tests\Fixtures\TypedProperties\Collection\Details\WithProductDescriptionTrait;
-
-trait WithTraitInsideTrait
-{
-    use WithProductDescriptionTrait;
-}
